move cobra commands to musicworld package and check for musicworld_dev container with docker inspect before attempting to exec into it
